refactor: vc download methods to a file
