+++ sample-application/sample-application/app/src/main/java/com/example/samplecredentialwallet/utils/CredentialStore.kt
